+++ laonog-modules/laonog-admin-service/src/main/java/com/laonog/admin/model/dto/TreeNode.java
@@ -10,8 +10,8 @@ import java.util.List;
  */
 @Data
 public class TreeNode {
-    protected int id;
-    protected int parentId;
+    protected long id;
+    protected long parentId;
     protected List<TreeNode> children = new ArrayList<TreeNode>();
 
     public void add(TreeNode node) {
